#109 extracted DatabaseManager interface - extracted interface - renamed DatabaseManager class - added stub for delete and update methods
